fix(api): move StepTaker to it's own pkg
