+++ images/virtualization-artifact/pkg/common/steptaker/runner.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package blockdevice
+package steptaker
 
 import (
 	"context"
